+++ src/js/dataMapping/navigation/menuOptions.jsx
@@ -47,13 +47,7 @@ export const profileOptions = [
         enabled: true
     }
 ];
-const federalSpendingGuide = () => {
-    if (window.location.href === 'http://localhost:3000/analyst-guide') {
-        return window.location.replace('http://localhost:3000/federal-spending-guide');
-    }
-    return '/federal-spending-guide';
-};
-console.log("here", federalSpendingGuide());
+
 export const resourceOptions = [
     {
         label: 'Glossary',
@@ -91,7 +85,7 @@ export const resourceOptions = [
     {
         label: 'Federal Spending Guide',
         type: 'analyst-guide',
-        url: `${federalSpendingGuide()}`,
+        url: '/federal-spending-guide',
         shouldOpenNewTab: false,
         enabled: true,
         externalLink: false
